Split IPv4 fragmentation tests to a separate module
The split tests module makes it easier to avoid using the internals of the
"apps.lwaftr.fragmentv4" module while testing it, and keeps the code of the
latter cleaner.

As a bonus, the new "apps.lwaftr.fragmentv4_test" module contains guards
which allow it to be both imported via "require()" (e.g. to call its
"selftest()" function from another test harness), and to be run directly
with "snsh":

   sudo ./snabb snsh ./apps/lwaftr/fragmentv4_test.lua
